Refactoring history service and update roadmap
